Now Welcomes the right user in account.php Added a new block for active users information in cards.php Added constrictions for displaying buttons in serie.php Defined $display_buttons on serie 'get' page, added login checks on several pages, removed redundant code, added error messages from POST route to some pages in index.php Updated several functions, Added $active_user_id variable to add_serie(), update_serie() and remove_serie() functions, added docstrings to newly created functions in model.php

+++ week2/model.php
@@ -228,6 +228,9 @@ function get_error($feedback){
  * @return array with message feedback
  */
 function add_serie($pdo, $serie_info){
+    /* Get active user id */
+    $active_user_id = $_SESSION['user_id'];
+
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -259,14 +262,14 @@ function add_serie($pdo, $serie_info){
             'message' => 'This series was already added.'
         ];
     }
-
     /* Add Serie */
-    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract) VALUES (?, ?, ?, ?)");
+    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract, user) VALUES (?, ?, ?, ?, ?)");
     $stmt->execute([
         $serie_info['Name'],
         $serie_info['Creator'],
         $serie_info['Seasons'],
-        $serie_info['Abstract']
+        $serie_info['Abstract'],
+        $active_user_id
     ]);
     $inserted = $stmt->rowCount();
     if ($inserted ==  1) {
@@ -290,66 +293,75 @@ function add_serie($pdo, $serie_info){
  * @return array
  */
 function update_serie($pdo, $serie_info){
-    /* Check if all fields are set */
-    if (
-        empty($serie_info['Name']) or
-        empty($serie_info['Creator']) or
-        empty($serie_info['Seasons']) or
-        empty($serie_info['Abstract']) or
-        empty($serie_info['serie_id'])
-    ) {
-        return [
-            'type' => 'danger',
-            'message' => 'There was an error. Not all fields were filled in.'
-        ];
-    }
+    /* Get active user id */
+    $active_user_id = $_SESSION['user_id'];
 
-    /* Check data type */
-    if (!is_numeric($serie_info['Seasons'])) {
-        return [
-            'type' => 'danger',
-            'message' => 'There was an error. You should enter a number in the field Seasons.'
-        ];
-    }
-
-    /* Get current series name */
+    /* Get id of user who added the serie */
     $stmt = $pdo->prepare('SELECT * FROM series WHERE id = ?');
     $stmt->execute([$serie_info['serie_id']]);
     $serie = $stmt->fetch();
-    $current_name = $serie['name'];
+    $serie_user_id = $serie['user'];
 
-    /* Check if serie already exists */
-    $stmt = $pdo->prepare('SELECT * FROM series WHERE name = ?');
-    $stmt->execute([$serie_info['Name']]);
-    $serie = $stmt->fetch();
-    if ($serie_info['Name'] == $serie['name'] and $serie['name'] != $current_name){
-        return [
-            'type' => 'danger',
-            'message' => sprintf("The name of the series cannot be changed. %s already exists.", $serie_info['Name'])
-        ];
-    }
+    if ($active_user_id == $serie_user_id) {
+        /* Check if all fields are set */
+        if (
+            empty($serie_info['Name']) or
+            empty($serie_info['Creator']) or
+            empty($serie_info['Seasons']) or
+            empty($serie_info['Abstract']) or
+            empty($serie_info['serie_id'])
+        ) {
+            return [
+                'type' => 'danger',
+                'message' => 'There was an error. Not all fields were filled in.'
+            ];
+        }
 
-    /* Update Serie */
-    $stmt = $pdo->prepare("UPDATE series SET name = ?, creator = ?, seasons = ?, abstract = ? WHERE id = ?");
-    $stmt->execute([
-        $serie_info['Name'],
-        $serie_info['Creator'],
-        $serie_info['Seasons'],
-        $serie_info['Abstract'],
-        $serie_info['serie_id']
-    ]);
-    $updated = $stmt->rowCount();
-    if ($updated ==  1) {
-        return [
-            'type' => 'success',
-            'message' => sprintf("Series '%s' was edited!", $serie_info['Name'])
-        ];
-    }
-    else {
-        return [
-            'type' => 'warning',
-            'message' => 'The series was not edited. No changes were detected'
-        ];
+        /* Check data type */
+        if (!is_numeric($serie_info['Seasons'])) {
+            return [
+                'type' => 'danger',
+                'message' => 'There was an error. You should enter a number in the field Seasons.'
+            ];
+        }
+
+        /* Get current series name */
+        $stmt = $pdo->prepare('SELECT * FROM series WHERE id = ?');
+        $stmt->execute([$serie_info['serie_id']]);
+        $serie = $stmt->fetch();
+        $current_name = $serie['name'];
+
+        /* Check if serie already exists */
+        $stmt = $pdo->prepare('SELECT * FROM series WHERE name = ?');
+        $stmt->execute([$serie_info['Name']]);
+        $serie = $stmt->fetch();
+        if ($serie_info['Name'] == $serie['name'] and $serie['name'] != $current_name) {
+            return [
+                'type' => 'danger',
+                'message' => sprintf("The name of the series cannot be changed. %s already exists.", $serie_info['Name'])
+            ];
+        }
+        /* Update Serie */
+        $stmt = $pdo->prepare("UPDATE series SET name = ?, creator = ?, seasons = ?, abstract = ? WHERE id = ?");
+        $stmt->execute([
+            $serie_info['Name'],
+            $serie_info['Creator'],
+            $serie_info['Seasons'],
+            $serie_info['Abstract'],
+            $serie_info['serie_id']
+        ]);
+        $updated = $stmt->rowCount();
+        if ($updated == 1) {
+            return [
+                'type' => 'success',
+                'message' => sprintf("Series '%s' was edited!", $serie_info['Name'])
+            ];
+        } else {
+            return [
+                'type' => 'warning',
+                'message' => 'The series was not edited. No changes were detected'
+            ];
+        }
     }
 }
 
@@ -360,24 +372,32 @@ function update_serie($pdo, $serie_info){
  * @return array
  */
 function remove_serie($pdo, $serie_id){
-    /* Get series info */
-    $serie_info = get_serieinfo($pdo, $serie_id);
+    /* Get active user id */
+    $active_user_id = $_SESSION['user_id'];
 
-    /* Delete Serie */
-    $stmt = $pdo->prepare("DELETE FROM series WHERE id = ?");
-    $stmt->execute([$serie_id]);
-    $deleted = $stmt->rowCount();
-    if ($deleted ==  1) {
-        return [
-            'type' => 'success',
-            'message' => sprintf("Series '%s' was removed!", $serie_info['name'])
-        ];
-    }
-    else {
-        return [
-            'type' => 'warning',
-            'message' => 'An error occurred. The series was not removed.'
-        ];
+    /* Get id of user who added the serie */
+    $serie_info = get_serieinfo($pdo, $serie_id);
+    $stmt = $pdo->prepare('SELECT * FROM series WHERE id = ?');
+    $stmt->execute([$serie_info['serie_id']]);
+    $serie = $stmt->fetch();
+    $serie_user_id = $serie['user'];
+
+    if ($active_user_id == $serie_user_id) {
+        /* Delete Serie */
+        $stmt = $pdo->prepare("DELETE FROM series WHERE id = ?");
+        $stmt->execute([$serie_id]);
+        $deleted = $stmt->rowCount();
+        if ($deleted == 1) {
+            return [
+                'type' => 'success',
+                'message' => sprintf("Series '%s' was removed!", $serie_info['name'])
+            ];
+        } else {
+            return [
+                'type' => 'warning',
+                'message' => 'An error occurred. The series was not removed.'
+            ];
+        }
     }
 }
 
@@ -416,6 +436,12 @@ function get_user_id(){
     }
 }
 
+/**
+ * Retrieves the first name and last name of a user from the database.
+ * @param $user_id
+ * @param $pdo
+ * @return string of first name and last name of user
+ */
 function get_user_name($user_id, $pdo) {
     $stmt = $pdo->prepare('SELECT firstname, lastname FROM users WHERE id = ?');
     $stmt->execute([$user_id]);
@@ -430,10 +456,155 @@ function get_user_name($user_id, $pdo) {
     return "$first_name $last_name";
 }
 
+/**
+ * Returns a count of the amount of users that are signed up
+ * @param $pdo
+ * @return mixed
+ */
 function count_users($pdo){
     /* Get series */
     $stmt = $pdo->prepare('SELECT * FROM users');
     $stmt->execute();
     $users = $stmt->rowCount();
     return $users;
+}
+
+/**
+ * Allows registering an account
+ * @param $pdo
+ * @param $form_data
+ * @return array
+ */
+function register_user($pdo, $form_data) {
+    if (
+        empty($form_data['username']) or
+        empty($form_data['password']) or
+        empty($form_data['firstname']) or
+        empty($form_data['lastname'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'You should enter a username, password, first- and last name.'
+        ];
+    }
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_exists = $stmt->rowCount();
+    } catch (\PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /* Return error message for existing username */
+    if ( !empty($user_exists) ) {
+        return [
+            'type' => 'danger',
+            'message' => 'The username you entered does already exist!'
+        ];
+    }
+    /* Hash password */
+    $password = password_hash($form_data['password'], PASSWORD_DEFAULT);
+    /* Save user to the database */
+    try {
+        $stmt = $pdo->prepare('INSERT INTO users (username, password, firstname, lastname) VALUES (?, ?, ?, ?)');
+        $stmt->execute([$form_data['username'], $password, $form_data['firstname'],
+            $form_data['lastname']]);
+        $user_id = $pdo->lastInsertId();
+    } catch (PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /* Login user and redirect */
+    session_start();
+    $_SESSION['user_id'] = $user_id;
+    $feedback = [
+        'type' => 'success',
+        'message' => sprintf('%s, your account was successfully created!', get_user_name($_SESSION['user_id'], $pdo))
+    ];
+    redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s',
+        json_encode($feedback)));
+
+}
+
+/**
+ * Allows logging in
+ * @param $pdo
+ * @param $form_data
+ * @return array
+ */
+function login_user($pdo, $form_data) {
+    echo $form_data;
+    /* Check if all fields are set */
+    if (
+        empty($form_data['username']) or
+        empty($form_data['password'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'You should enter a username and password.'
+        ];
+    }
+    /* Check if user exists */
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_info = $stmt->fetch();
+    } catch (\PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /* Return error message for wrong username */
+    if (empty($user_info)) {
+        return [
+            'type' => 'danger',
+            'message' => 'The username you entered does not exist!'
+        ];
+    }
+    /* Check password */
+    echo $form_data['username'];
+    echo $form_data['password'];
+    if (!password_verify($form_data['password'], $user_info['password'])) {
+        return [
+            'type' => 'danger',
+            'message' => 'The password you entered is incorrect!'
+        ];
+    } else {
+        session_start();
+        $_SESSION['user_id'] = $user_info['id'];
+        $feedback = [
+            'type' => 'success',
+            'message' => sprintf('%s, you were logged in successfully!',
+                get_user_name($_SESSION['user_id'], $pdo))
+        ];
+        redirect(sprintf('/DDWT18/week2/myaccount/?error_msg=%s',
+            json_encode($feedback)));
+    }
+}
+
+/**
+ * Returns true when a user is logged in.
+ * @return bool
+ */
+function check_login(){
+    session_start();
+    if (isset($_SESSION['user_id'])){
+        return True;
+    } else {
+        return False;
+    }
+}
+
+/**
+ * Logs out a user.
+ */
+function logout_user() {
+    session_start() ;
+    session_destroy() ;
+    $_SESSION = array();
 }
